+++ src/cloregram/impl/validation/state.clj
@@ -1,4 +1,4 @@
-(ns cloregram.test.infrastructure.state)
+(ns cloregram.impl.validation.state)
 
 (defonce webhook-address (atom nil))
 
@@ -6,6 +6,8 @@
 
 (defonce update-id (atom 0))
 
-(defonce users (atom {}))
+(defonce v-users (atom {}))
 
 (defonce checkout-queries (atom {}))
+
+(defonce files (atom {}))