+++ src/cloregram/schema.clj
@@ -1,5 +1,0 @@
-(ns cloregram.schema
-  (:require [cloregram.schema.user :refer [user]]
-            [cloregram.schema.callback :refer [callback]]))
-
-(def schema (concat user callback))