+++ src/cloregram/utils.clj
@@ -1,35 +1,10 @@
 (ns cloregram.utils
-  (:require [clojure.java.io :as io]
-            [clojure.edn :as edn]
-            [clojure.string :as str]
-            [resauce.core :as res]
-            [taoensso.timbre :as log]))
-
-(defn dbg
-  ([x] (dbg nil x))
-  ([msg x]
-   (log/debug msg x)
-   x))
-
-(defn api-wrap
-  [api-f & args]
-  (log/debug "API method calling" {:method api-f
-                                   :arguments args})
-  (let [resp  (apply api-f args)
-        ok    (-> resp :ok true?)
-        desc  (:description resp)]
-    (when (not ok)
-      (throw (ex-info "API response error" {:method api-f
-                                            :arguments args
-                                            :error desc
-                                            :response resp})))
-    (log/debug "Response is OK" {:method api-f
-                                 :arguments args
-                                 :response resp})
-    (:result resp)))
+  (:require [com.brunobonacci.mulog :as μ]))
 
 (defn deep-merge
+  
   "Recursively merges maps"
+
   [& maps]
   (letfn [(m [& xs]
             (if (some #(and (map? %) (not (record? %))) xs)
@@ -37,51 +12,45 @@
               (last xs)))]
     (reduce m maps)))
 
-(defn keys-hyphens->underscores ; NOT recursive!
-  [m]
-  (into {} (map (fn [[k v]] [(-> k name (.replace \- \_) keyword) v]) m)))
-
-(defn simplify-reply-markup
-  [reply-markup]
-  (vec (map #(vec (map :text %)) reply-markup)))
-
-(defn msg->str
-  [msg]
-  (let [msg (-> msg
-                (select-keys [:text :reply_markup])
-                (update :reply_markup simplify-reply-markup))]
-    (format "%s\t%s" (:text msg) (:reply_markup msg))))
-
 (defn username
+
+  "Helper to get human readable identificator of `user` even he don't have Telegram username"
+  
   [user]
   (or (:user/username user) (str "id" (:user/id user))))
 
 (defmacro get-project-info
+
+  "This macro expands in map with keys `group`, `name` and `version` of current project by information from project.clj"
+  
   []
-  (let [[_ ga version] (read-string (slurp "project.clj"))
+  (let [[_ ga version] (read-string (try (slurp "project.clj") (catch Exception e "")))
         [ns name version] [(namespace ga) (name ga) version]]
     {:group ns
      :name name
      :version version}))
 
 (defn resolver
+
+  "Resolves symbol to value if exists"
+
   [sym]
   (let [ns (-> sym namespace symbol)
         nm (-> sym name symbol)]
     (require ns)
     (if-let [resolved (ns-resolve ns nm)]
       resolved
-      (log/error "Callback not resolved" {:callback-symbol sym}))))
+      (μ/log ::symbol-not-resolved :symbol sym))))
 
-(defn- read-resource [resource-url]
-  (with-open [stream (io/input-stream resource-url)]
-    (-> stream
-        io/reader
-        java.io.PushbackReader. edn/read)))
+(defn img-comparer
+  [img1 img2]
+  (let [h (.getHeight img1)
+        w (.getWidth img1)]
+    (loop [[x y] [0 0]]
+      (cond
+        (= y h) true
+        (not= (.getRGB img1 x y) (.getRGB img2 x y)) false
+        :else (let [x' (if (= (dec w) x) 0 (inc x))
+                    y' (if (= 0 x) (inc y) y)]
+                (recur [x' y']))))))
 
-(defn read-resource-dir
-  [dir]
-  (when-let [resources (some-> dir io/resource res/url-dir)]
-    (->> resources
-         (filter #(str/ends-with? % ".edn"))
-         (mapcat read-resource))))